+++ addons/mrp/static/src/components/bom_overview_line/mrp_bom_overview_line.js
@@ -2,7 +2,8 @@
 
 import { _t } from "@web/core/l10n/translation";
 import { useService } from "@web/core/utils/hooks";
-import { formatFloat, formatFloatTime, formatMonetary } from "@web/views/fields/formatters";
+import { formatFloatTime, formatMonetary } from "@web/views/fields/formatters";
+import { formatFloat } from "@web/core/utils/numbers";
 
 const { Component } = owl;
