+++ addons/mrp/static/src/components/bom_overview_table/mrp_bom_overview_table.js
@@ -1,6 +1,7 @@
 /** @odoo-module **/
 
-import { formatMonetary, formatFloat } from "@web/views/fields/formatters";
+import { formatMonetary } from "@web/views/fields/formatters";
+import { formatFloat } from "@web/core/utils/numbers";
 import { useService } from "@web/core/utils/hooks";
 import { BomOverviewLine } from "../bom_overview_line/mrp_bom_overview_line";
 import { BomOverviewComponentsBlock } from "../bom_overview_components_block/mrp_bom_overview_components_block";
