+++ addons/pos_loyalty/static/src/overrides/components/partner_line/partner_line.js
@@ -4,7 +4,7 @@ import { _t } from "@web/core/l10n/translation";
 import { usePos } from "@point_of_sale/app/store/pos_hook";
 import { PartnerLine } from "@point_of_sale/app/screens/partner_list/partner_line/partner_line";
 import { patch } from "@web/core/utils/patch";
-import { formatFloat } from "@web/views/fields/formatters";
+import { formatFloat } from "@web/core/utils/numbers";
 
 patch(PartnerLine.prototype, {
     setup() {
