+++ addons/sale_product_configurator/static/src/js/badge_extra_price/badge_extra_price.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
 import { Component } from "@odoo/owl";
-import { formatMonetary } from "@web/views/fields/formatters";
+import { formatCurrency } from "@web/core/currency";
 
 export class BadgeExtraPrice extends Component {
     static template = "product.badge_extra_price";
@@ -16,6 +16,6 @@ export class BadgeExtraPrice extends Component {
      * @return {String} - The price, in the format of the given currency.
      */
     getFormattedPrice() {
-        return formatMonetary( Math.abs(this.props.price), {currencyId: this.props.currencyId});
+        return formatCurrency( Math.abs(this.props.price), this.props.currencyId);
     }
 }
