+++ addons/stock/static/src/components/reception_report_line/stock_reception_report_line.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 import { useService } from "@web/core/utils/hooks";
-import { formatFloat } from "@web/views/fields/formatters";
+import { formatFloat } from "@web/core/utils/numbers";
 
 const { Component } = owl;
