+++ addons/stock/static/src/stock_forecasted/forecasted_details.js
@@ -1,5 +1,5 @@
 /** @odoo-module **/
-import { formatFloat } from "@web/views/fields/formatters";
+import { formatFloat } from "@web/core/utils/numbers";
 import { useService } from "@web/core/utils/hooks";
 
 const { Component } = owl;
