+++ addons/web/static/src/views/fields/gauge/gauge_field.js
@@ -3,7 +3,7 @@
 import { _t } from "@web/core/l10n/translation";
 import { loadJS } from "@web/core/assets";
 import { registry } from "@web/core/registry";
-import { formatFloat } from "@web/views/fields/formatters";
+import { formatFloat } from "@web/core/utils/numbers";
 import { standardFieldProps } from "@web/views/fields/standard_field_props";
 
 import { Component, onWillStart, useEffect, useRef } from "@odoo/owl";
