+++ addons/web/static/tests/core/utils/numbers_tests.js
@@ -1,6 +1,10 @@
 /** @odoo-module **/
 
-import { roundPrecision, roundDecimals, floatIsZero } from "@web/core/utils/numbers";
+import { defaultLocalization } from "@web/../tests/helpers/mock_services";
+import { patchWithCleanup } from "@web/../tests/helpers/utils";
+import { localization } from "@web/core/l10n/localization";
+
+import { roundPrecision, roundDecimals, floatIsZero, formatFloat } from "@web/core/utils/numbers";
 
 QUnit.module("utils", () => {
     QUnit.module("numbers");
@@ -135,5 +139,145 @@ QUnit.module("utils", () => {
         assert.strictEqual(floatIsZero(-0.00000005, 7), false);
         assert.strictEqual(floatIsZero(-0.000000099999, 7), false);
         assert.strictEqual(floatIsZero(-0.0000001, 7), false);
+    });
+});
+
+QUnit.module("utils", (hooks) => {
+    hooks.beforeEach(() => {
+        patchWithCleanup(localization, { ...defaultLocalization, grouping: [3, 0] });
+    });
+
+    QUnit.module("numbers");
+
+    QUnit.test("formatFloat", function (assert) {
+        assert.strictEqual(formatFloat(1000000), "1,000,000.00");
+
+        const options = { grouping: [3, 2, -1], decimalPoint: "?", thousandsSep: "€" };
+        assert.strictEqual(formatFloat(106500, options), "1€06€500?00");
+
+        assert.strictEqual(formatFloat(1500, { thousandsSep: "" }), "1500.00");
+        assert.strictEqual(formatFloat(-1.01), "-1.01");
+        assert.strictEqual(formatFloat(-0.01), "-0.01");
+
+        assert.strictEqual(formatFloat(38.0001, { trailingZeros: false }), "38");
+        assert.strictEqual(formatFloat(38.1, { trailingZeros: false }), "38.1");
+        assert.strictEqual(formatFloat(38.0001, { digits: [16, 0], trailingZeros: false }), "38");
+
+        patchWithCleanup(localization, { grouping: [3, 3, 3, 3] });
+        assert.strictEqual(formatFloat(1000000), "1,000,000.00");
+
+        patchWithCleanup(localization, { grouping: [3, 2, -1] });
+        assert.strictEqual(formatFloat(106500), "1,06,500.00");
+
+        patchWithCleanup(localization, { grouping: [1, 2, -1] });
+        assert.strictEqual(formatFloat(106500), "106,50,0.00");
+
+        patchWithCleanup(localization, {
+            grouping: [2, 0],
+            decimalPoint: "!",
+            thousandsSep: "@",
+        });
+        assert.strictEqual(formatFloat(6000), "60@00!00");
+    });
+
+    QUnit.test("formatFloat (humanReadable=true)", async (assert) => {
+        assert.strictEqual(
+            formatFloat(1020, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "1.02k"
+        );
+        assert.strictEqual(
+            formatFloat(1020000, { humanReadable: true, decimals: 2, minDigits: 2 }),
+            "1,020k"
+        );
+        assert.strictEqual(
+            formatFloat(10200000, { humanReadable: true, decimals: 2, minDigits: 2 }),
+            "10.20M"
+        );
+        assert.strictEqual(
+            formatFloat(1020, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "1.02k"
+        );
+        assert.strictEqual(
+            formatFloat(1002, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "1.00k"
+        );
+        assert.strictEqual(
+            formatFloat(101, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "101.00"
+        );
+        assert.strictEqual(
+            formatFloat(64.2, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "64.20"
+        );
+        assert.strictEqual(formatFloat(1e18, { humanReadable: true }), "1E");
+        assert.strictEqual(
+            formatFloat(1e21, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "1e+21"
+        );
+        assert.strictEqual(
+            formatFloat(1.0045e22, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "1e+22"
+        );
+        assert.strictEqual(
+            formatFloat(1.0045e22, { humanReadable: true, decimals: 3, minDigits: 1 }),
+            "1.005e+22"
+        );
+        assert.strictEqual(
+            formatFloat(1.012e43, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "1.01e+43"
+        );
+        assert.strictEqual(
+            formatFloat(1.012e43, { humanReadable: true, decimals: 2, minDigits: 2 }),
+            "1.01e+43"
+        );
+        assert.strictEqual(
+            formatFloat(-1020, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-1.02k"
+        );
+        assert.strictEqual(
+            formatFloat(-1020000, { humanReadable: true, decimals: 2, minDigits: 2 }),
+            "-1,020k"
+        );
+        assert.strictEqual(
+            formatFloat(-10200000, { humanReadable: true, decimals: 2, minDigits: 2 }),
+            "-10.20M"
+        );
+        assert.strictEqual(
+            formatFloat(-1020, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-1.02k"
+        );
+        assert.strictEqual(
+            formatFloat(-1002, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-1.00k"
+        );
+        assert.strictEqual(
+            formatFloat(-101, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-101.00"
+        );
+        assert.strictEqual(
+            formatFloat(-64.2, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-64.20"
+        );
+        assert.strictEqual(formatFloat(-1e18, { humanReadable: true }), "-1E");
+        assert.strictEqual(
+            formatFloat(-1e21, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-1e+21"
+        );
+        assert.strictEqual(
+            formatFloat(-1.0045e22, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-1e+22"
+        );
+        assert.strictEqual(
+            formatFloat(-1.0045e22, { humanReadable: true, decimals: 3, minDigits: 1 }),
+            "-1.004e+22"
+        );
+        assert.strictEqual(
+            formatFloat(-1.012e43, { humanReadable: true, decimals: 2, minDigits: 1 }),
+            "-1.01e+43"
+        );
+        assert.strictEqual(
+            formatFloat(-1.012e43, { humanReadable: true, decimals: 2, minDigits: 2 }),
+            "-1.01e+43"
+        );
     });
 });
